Bio.Parse.Sequence.FastQ

Documentation

data FastQSequence

Constructors

FastQSequence

Fields

_header :: [ByteString]

_sequence :: ByteString

_quality :: ByteString

sequence :: Lens' FastQSequence ByteString

quality :: Lens' FastQSequence ByteString

header :: Lens' FastQSequence [ByteString]

parseHeader :: ParseConstraint m => m [ByteString]

Parses the header of a FASTQ file.

The header starts with a @ character and follows with an optional description. When making use if BioSeq, the first word of the header is used as the seqid.

parseSequenceLine :: ParseConstraint m => m ByteString

Parses the line of sequence data.

parseQualityLine :: ParseConstraint m => m ByteString

Parses the line of quality data.

parseSequence' :: ParseConstraint m => m FastQSequence

Parses an entire sequence including its header and quality data.

parseSequences' :: ParseConstraint m => m [FastQSequence]

Parses many sequences.