Bio.Parse.Sequence.FastQ

Documentation

data FastQSequence

Constructors

FastQSequence

Fields

_header :: [ByteString]

_sequence :: ByteString

_quality :: ByteString

sequence :: Lens' FastQSequence ByteString

quality :: Lens' FastQSequence ByteString

header :: Lens' FastQSequence [ByteString]

parseHeader :: Parser [ByteString]

Parses the header of a FASTQ file.

The header starts with a @ character and follows with an optional description. When making use if BioSeq, the first word of the header is used as the seqid.

parseSequenceLine :: Parser ByteString

Parses the line of sequence data.

parseQualityLine :: Parser ByteString

Parses the line of quality data.

parseSequence :: Parser FastQSequence

Parses an entire sequence including its header and quality data.

parseSequences :: Parser [FastQSequence]

Parses many sequences.

parseFastQ :: String -> Result [FastQSequence]

Parses sequences from a String.

parseFastQ x = parseString parseSequences mempty x

parseFastQB :: ByteString -> Result [FastQSequence]

Parses sequences from a strict ByteString.

parseFastQB x = parseByteString parseSequences mempty x